+++ server/src/main/java/org/opensearch/index/IndexSettings.java
@@ -950,6 +950,8 @@ public final class IndexSettings {
      */
     private final boolean isCompositeIndex;
 
+    private boolean isRemoteStoreSSEnabled;
+
     /**
      * Denotes whether search via star tree index is enabled for this index
      */
@@ -1035,6 +1037,7 @@ public final class IndexSettings {
         numberOfShards = settings.getAsInt(IndexMetadata.SETTING_NUMBER_OF_SHARDS, null);
         replicationType = IndexMetadata.INDEX_REPLICATION_TYPE_SETTING.get(settings);
         isRemoteStoreEnabled = settings.getAsBoolean(IndexMetadata.SETTING_REMOTE_STORE_ENABLED, false);
+        isRemoteStoreSSEnabled = settings.getAsBoolean(IndexMetadata.SETTING_REMOTE_STORE_SSE_ENABLED, false);
 
         isWarmIndex = settings.getAsBoolean(IndexModule.IS_WARM_INDEX_SETTING.getKey(), false);
 
@@ -1239,6 +1242,9 @@ public final class IndexSettings {
         );
         scopedSettings.addSettingsUpdateConsumer(ALLOW_DERIVED_FIELDS, this::setAllowDerivedField);
         scopedSettings.addSettingsUpdateConsumer(IndexMetadata.INDEX_REMOTE_STORE_ENABLED_SETTING, this::setRemoteStoreEnabled);
+
+        scopedSettings.addSettingsUpdateConsumer(IndexMetadata.INDEX_REMOTE_STORE_SSE_ENABLED_SETTING, this::setRemoteStoreSseEnabled);
+
         scopedSettings.addSettingsUpdateConsumer(
             IndexMetadata.INDEX_REMOTE_SEGMENT_STORE_REPOSITORY_SETTING,
             this::setRemoteStoreRepository
@@ -1425,6 +1431,13 @@ public final class IndexSettings {
      */
     public boolean isRemoteStoreEnabled() {
         return isRemoteStoreEnabled;
+    }
+
+    /**
+     * Returns if remote store is enabled for this index.
+     */
+    public boolean isRemoteStoreSSEnabled() {
+        return isRemoteStoreSSEnabled;
     }
 
     public boolean isAssignedOnRemoteNode() {
@@ -2135,6 +2148,10 @@ public final class IndexSettings {
 
     public void setRemoteStoreEnabled(boolean isRemoteStoreEnabled) {
         this.isRemoteStoreEnabled = isRemoteStoreEnabled;
+    }
+
+    public void setRemoteStoreSseEnabled(boolean sseEnabled) {
+        this.isRemoteStoreSSEnabled = sseEnabled;
     }
 
     public void setRemoteStoreRepository(String remoteStoreRepository) {