+++ server/src/test/java/org/opensearch/index/translog/RemoteFsTranslogTests.java
@@ -234,7 +234,7 @@ public class RemoteFsTranslogTests extends OpenSearchTestCase {
             new ByteSizeValue(10 + randomInt(128 * 1024), ByteSizeUnit.BYTES)
         );
         // To simulate that the node is remote backed
-        Settings nodeSettings = Settings.builder().put("node.attr.remote_store.translog.repository", "my-repo-1").build();
+        Settings nodeSettings = Settings.builder().put("node.attr.remote_store.translog.repository", "tlog-test-repo").build();
         final IndexSettings indexSettings = IndexSettingsModule.newIndexSettings(shardId.getIndex(), settings, nodeSettings);
         return new TranslogConfig(shardId, path, indexSettings, NON_RECYCLING_INSTANCE, bufferSize, "", false);
     }