+++ test/framework/src/main/java/org/opensearch/index/shard/IndexShardTestUtils.java
@@ -18,6 +18,7 @@ import org.opensearch.node.remotestore.RemoteStoreNodeAttribute;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 
 public class IndexShardTestUtils {
@@ -41,7 +42,83 @@ public class IndexShardTestUtils {
         remoteNodeAttributes.put(RemoteStoreNodeAttribute.REMOTE_STORE_CLUSTER_STATE_REPOSITORY_NAME_ATTRIBUTE_KEY, MOCK_STATE_REPO_NAME);
         remoteNodeAttributes.put(RemoteStoreNodeAttribute.REMOTE_STORE_SEGMENT_REPOSITORY_NAME_ATTRIBUTE_KEY, MOCK_SEGMENT_REPO_NAME);
         remoteNodeAttributes.put(RemoteStoreNodeAttribute.REMOTE_STORE_TRANSLOG_REPOSITORY_NAME_ATTRIBUTE_KEY, MOCK_TLOG_REPO_NAME);
-        return new DiscoveryNode(
+        remoteNodeAttributes.put(
+            String.format(
+                Locale.getDefault(),
+                RemoteStoreNodeAttribute.REMOTE_STORE_REPOSITORY_TYPE_ATTRIBUTE_KEY_FORMAT,
+                MOCK_SEGMENT_REPO_NAME
+            ),
+            "s3"
+        );
+        remoteNodeAttributes.put(
+            String.format(
+                Locale.getDefault(),
+                RemoteStoreNodeAttribute.REMOTE_STORE_REPOSITORY_TYPE_ATTRIBUTE_KEY_FORMAT,
+                MOCK_TLOG_REPO_NAME
+            ),
+            "s3"
+        );
+        remoteNodeAttributes.put(
+            String.format(
+                Locale.getDefault(),
+                RemoteStoreNodeAttribute.REMOTE_STORE_REPOSITORY_TYPE_ATTRIBUTE_KEY_FORMAT,
+                MOCK_STATE_REPO_NAME
+            ),
+            "s3"
+        );
+
+        remoteNodeAttributes.put(
+            String.format(
+                Locale.getDefault(),
+                RemoteStoreNodeAttribute.REMOTE_STORE_REPOSITORY_SETTINGS_ATTRIBUTE_KEY_PREFIX,
+                MOCK_SEGMENT_REPO_NAME
+            ) + "bucket",
+            "test-bucket"
+        );
+        remoteNodeAttributes.put(
+            String.format(
+                Locale.getDefault(),
+                RemoteStoreNodeAttribute.REMOTE_STORE_REPOSITORY_SETTINGS_ATTRIBUTE_KEY_PREFIX,
+                MOCK_SEGMENT_REPO_NAME
+            ) + "path",
+            "test-path"
+        );
+
+        remoteNodeAttributes.put(
+            String.format(
+                Locale.getDefault(),
+                RemoteStoreNodeAttribute.REMOTE_STORE_REPOSITORY_SETTINGS_ATTRIBUTE_KEY_PREFIX,
+                MOCK_TLOG_REPO_NAME
+            ) + "bucket",
+            "test-bucket"
+        );
+        remoteNodeAttributes.put(
+            String.format(
+                Locale.getDefault(),
+                RemoteStoreNodeAttribute.REMOTE_STORE_REPOSITORY_SETTINGS_ATTRIBUTE_KEY_PREFIX,
+                MOCK_TLOG_REPO_NAME
+            ) + "path",
+            "test-path"
+        );
+
+        remoteNodeAttributes.put(
+            String.format(
+                Locale.getDefault(),
+                RemoteStoreNodeAttribute.REMOTE_STORE_REPOSITORY_SETTINGS_ATTRIBUTE_KEY_PREFIX,
+                MOCK_STATE_REPO_NAME
+            ) + "bucket",
+            "test-bucket"
+        );
+        remoteNodeAttributes.put(
+            String.format(
+                Locale.getDefault(),
+                RemoteStoreNodeAttribute.REMOTE_STORE_REPOSITORY_SETTINGS_ATTRIBUTE_KEY_PREFIX,
+                MOCK_STATE_REPO_NAME
+            ) + "path",
+            "test-path"
+        );
+
+        DiscoveryNode node = new DiscoveryNode(
             id,
             id,
             IndexShardTestCase.buildNewFakeTransportAddress(),
@@ -49,6 +126,9 @@ public class IndexShardTestUtils {
             DiscoveryNodeRole.BUILT_IN_ROLES,
             Version.CURRENT
         );
+        RemoteStoreNodeAttribute remoteStoreNodeAttribute = new RemoteStoreNodeAttribute(node);
+        assert (remoteStoreNodeAttribute.getRepositoriesMetadata() != null);
+        return node;
     }
 
     public static DiscoveryNodes getFakeDiscoveryNodes(List<ShardRouting> shardRoutings) {